pyproject.toml
[project]
name = "MediaPorter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.118.0",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.37.0",
    "yt-dlp>=2025.9.26",
]
